feature : #40 도메인 이름 오타 수정
