+++ src/main/java/com/mingles/metamingle/interactivemovie/command/domain/aggregate/vo/ShortFormNoVO.java
@@ -1,0 +1,21 @@
+package com.mingles.metamingle.interactivemovie.command.domain.aggregate.vo;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+
+@Getter
+@Embeddable
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ShortFormNoVO {
+
+    @Column
+    private Long shortFormNo;
+
+
+}
