+++ src/main/java/com/mingles/metamingle/interactivemovie/query/application/dto/request/temp.java
@@ -1,0 +1,4 @@
+package com.mingles.metamingle.interactivemovie.query.application.dto.request;
+
+public class temp {
+}
