+++ src/main/java/com/mingles/metamingle/interactivemovie/query/domain/service/temp.java
@@ -1,0 +1,4 @@
+package com.mingles.metamingle.interactivemovie.query.domain.service;
+
+public class temp {
+}
