+++ src/main/java/com/mingles/metamingle/interactivemovie/query/infrastructure/repository/temp.java
@@ -1,0 +1,5 @@
+
+package com.mingles.metamingle.interactivemovie.query.infrastructure.repository;
+
+public class temp {
+}
