+++ src/main/java/com/mingles/metamingle/interactivemovie/query/infrastructure/service/temp.java
@@ -1,0 +1,4 @@
+package com.mingles.metamingle.interactivemovie.query.infrastructure.service;
+
+public class temp {
+}
