--- src/main/java/com/mingles/metamingle/interativemovie/command/application/dto/VideoData.java
@@ -1,13 +1,0 @@
-package com.mingles.metamingle.interativemovie.command.application.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
-
-@Getter
-@Setter
-public class VideoData {
-    private List<String> files;
-    private VideoMetadata videoMetadata;
-}
