--- src/main/java/com/mingles/metamingle/interativemovie/command/application/dto/request/temp.java
@@ -1,4 +1,0 @@
-package com.mingles.metamingle.interativemovie.command.application.dto.request;
-
-public class temp {
-}
