--- src/main/java/com/mingles/metamingle/interativemovie/command/application/dto/response/CreateInteractiveMovieResponse.java
@@ -1,19 +1,0 @@
-package com.mingles.metamingle.interativemovie.command.application.dto.response;
-
-import lombok.*;
-
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class CreateInteractiveMovieResponse {
-
-    private Long shortFormNo;
-    private Long interactiveMovieNo;
-    private String thumbnailUrl;
-    private String url;
-    private String choice;
-    private int sequence;
-
-}
