--- src/main/java/com/mingles/metamingle/interativemovie/command/infrastructure/repository/temp.java
@@ -1,5 +1,0 @@
-
-package com.mingles.metamingle.interativemovie.command.infrastructure.repository;
-
-public class temp {
-}
