--- src/main/java/com/mingles/metamingle/interativemovie/query/domain/service/temp.java
@@ -1,4 +1,0 @@
-package com.mingles.metamingle.interativemovie.query.domain.service;
-
-public class temp {
-}
