--- src/main/java/com/mingles/metamingle/interativemovie/query/infrastructure/repository/temp.java
@@ -1,5 +1,0 @@
-
-package com.mingles.metamingle.interativemovie.query.infrastructure.repository;
-
-public class temp {
-}
