--- src/main/java/com/mingles/metamingle/interativemovie/query/infrastructure/service/temp.java
@@ -1,4 +1,0 @@
-package com.mingles.metamingle.interativemovie.query.infrastructure.service;
-
-public class temp {
-}
